requirements.txt
google-genai
chromadb
playwright
mcp
pyyaml
